Fix compile on BH by eliminating name rebind, exposing coordinate functions, and adding ipole machine-specific files
