+++ machines/bh-cluster.make
@@ -1,0 +1,1 @@
+CFLAGS_CUSTOM = -D_DEFAULT_SOURCE
