+++ machines/bh.make
@@ -1,0 +1,1 @@
+CFLAGS_CUSTOM = -shlib
